backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.35
PyMySQL==1.1.1
python-multipart==0.0.9
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
